@@ -10,31 +10,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.10",
-    "@angular/common": "^16.2.10",
-    "@angular/compiler": "^16.2.10",
-    "@angular/core": "^16.2.10",
-    "@angular/forms": "^16.2.10",
-    "@angular/platform-browser": "^16.2.10",
-    "@angular/platform-browser-dynamic": "^16.2.10",
-    "@angular/router": "^16.2.10",
-    "@ngrx/effects": "^16.3.0",
-    "@ngrx/store": "^16.3.0",
+    "@angular/animations": "^17.0.1",
+    "@angular/common": "^17.0.1",
+    "@angular/compiler": "^17.0.1",
+    "@angular/core": "^17.0.1",
+    "@angular/forms": "^17.0.1",
+    "@angular/platform-browser": "^17.0.1",
+    "@angular/platform-browser-dynamic": "^17.0.1",
+    "@angular/router": "^17.0.1",
+    "@ngrx/effects": "^17.0.0-beta.0",
+    "@ngrx/store": "^17.0.0-beta.0",
     "bootstrap": "^5.3.2",
     "jasmine-marbles": "^0.9.2",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.2",
-    "zone.js": "~0.13.2"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.7",
-    "@angular/cli": "^16.2.7",
-    "@angular/compiler-cli": "^16.2.10",
-    "@angular/language-service": "^16.2.10",
+    "@angular-devkit/build-angular": "^17.0.0",
+    "@angular/cli": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.1",
+    "@angular/language-service": "^17.0.1",
     "@chiragrupani/karma-chromium-edge-launcher": "^2.3.1",
-    "@types/jasmine": "~5.1.1",
-    "@types/jasminewd2": "~2.0.12",
-    "@types/node": "^20.8.7",
+    "@types/jasmine": "~5.1.2",
+    "@types/jasminewd2": "~2.0.13",
+    "@types/node": "^20.9.0",
     "codelyzer": "^6.0.2",
     "jasmine-core": "~5.1.1",
     "jasmine-spec-reporter": "~7.0.0",
